requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
billiard==4.2.0
cachetools==5.4.0
caldav==1.3.9
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
cryptography==43.0.0
defusedxml==0.8.0rc2
Django==5.0.8
django-admin-interface==0.28.8
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-colorfield==0.11.0
django-cors-headers==4.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.15.4
fire==0.6.0
fsspec==2024.6.1
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.24.5
icalendar==5.0.13
idna==3.7
inflection==0.5.1
Jinja2==3.1.4
kombu==5.4.0
llama_models==0.0.5
llama_toolchain==0.0.5
lxml==5.3.0
MarkupSafe==2.1.5
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
pyparsing==3.1.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.2
recurring-ical-events==2.2.3
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
termcolor==2.4.0
text-unidecode==1.3
tiktoken==0.7.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
vobject==0.9.7
wcwidth==0.2.13
x-wr-timezone==0.0.7
